Implemented RAG and added chat history using Langchain.
Not included in this checkin is modifying the frontend app to use the new langchain. Note that the enable_chat_history setting is ignored for now.

cloud_sql.py - now includes a method to create a PostgresEngine for storing and retrieving history, plus a CustomVectorStore to perform the query embedding and vector search.
rag_test.py - test script to demonstrate end-to-end RAG chain with chat history.
rag_chain.py - contains helper method create_chain to create the end-to-end RAG chain with history. Also contains a cache of chat history.

cloudbuild.yaml and Dockerfile.langchain - shows how to build the image off of the base image (not yet working)

various tf files: increased rag-frontend memory to 8g and cpu to 4g. increased max input and total tokens on HF TGI for mistral. threadded through some parameters needed to instantiate the PostgresEngine.